moved run fucntion into utils added protein option
